# valblk: fix Writer.Size() performance