@RaduBerinde 2024-12-18
I will backport to 24.3 (for when we add the cluster setting to enable value blocks).