@RaduBerinde 2024-12-18
I confirmed this is only an issue with `v5`:
```
BenchmarkWriterWithVersions/vals-per-key=10/format=(Pebble,v2)/block=32KB/filter=false/compression=NoCompression                      19          59592377 ns/op         660.38 MB/s
BenchmarkWriterWithVersions/vals-per-key=10/format=(Pebble,v3)/block=32KB/filter=false/compression=NoCompression                      12          94971892 ns/op         484.55 MB/s
BenchmarkWriterWithVersions/vals-per-key=10/format=(Pebble,v5)/block=32KB/filter=false/compression=NoCompression                       2         757362896 ns/op          52.63 MB/s
```